requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.2
    # via streamlit
celeste-ai==0.2.5
    # via celeste-staff-meal (pyproject.toml)
celeste-image-generation==0.2.1
    # via celeste-ai
celeste-image-intelligence==0.2.3
    # via
    #   celeste-staff-meal (pyproject.toml)
    #   celeste-ai
celeste-speech-generation==0.2.3
    # via celeste-ai
celeste-text-generation==0.2.1
    # via celeste-ai
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via streamlit
cryptography==46.0.3
    # via pyjwt
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
et-xmlfile==2.0.0
    # via openpyxl
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   celeste-ai
    #   postgrest
    #   storage3
    #   supabase
    #   supabase-auth
    #   supabase-functions
httpx-sse==0.4.3
    # via celeste-ai
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markupsafe==3.0.3
    # via jinja2
multidict==6.7.0
    # via yarl
narwhals==2.11.0
    # via
    #   altair
    #   plotly
numpy==2.3.4
    # via
    #   pandas
    #   pydeck
    #   streamlit
openpyxl==3.1.5
    # via celeste-staff-meal (pyproject.toml)
packaging==25.0
    # via
    #   altair
    #   deprecation
    #   plotly
    #   streamlit
pandas==2.3.3
    # via
    #   celeste-staff-meal (pyproject.toml)
    #   streamlit
pillow==12.0.0
    # via
    #   celeste-staff-meal (pyproject.toml)
    #   streamlit
plotly==6.4.0
    # via celeste-staff-meal (pyproject.toml)
postgrest==2.24.0
    # via supabase
propcache==0.4.1
    # via yarl
protobuf==6.33.1
    # via streamlit
pyarrow==21.0.0
    # via streamlit
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   celeste-staff-meal (pyproject.toml)
    #   celeste-ai
    #   postgrest
    #   pydantic-settings
    #   realtime
    #   storage3
    #   supabase-auth
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via
    #   celeste-staff-meal (pyproject.toml)
    #   celeste-ai
pydeck==0.9.1
    # via streamlit
pyjwt==2.10.1
    # via supabase-auth
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via
    #   celeste-staff-meal (pyproject.toml)
    #   celeste-ai
    #   pydantic-settings
pytz==2025.2
    # via pandas
qrcode==8.2
    # via celeste-staff-meal (pyproject.toml)
realtime==2.24.0
    # via supabase
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via streamlit
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
storage3==2.24.0
    # via supabase
streamlit==1.51.0
    # via celeste-staff-meal (pyproject.toml)
strenum==0.4.15
    # via supabase-functions
supabase==2.24.0
    # via celeste-staff-meal (pyproject.toml)
supabase-auth==2.24.0
    # via supabase
supabase-functions==2.24.0
    # via supabase
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   pydantic
    #   pydantic-core
    #   realtime
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
websockets==15.0.1
    # via realtime
yarl==1.22.0
    # via
    #   postgrest
    #   storage3
    #   supabase-functions
zxing-cpp==2.3.0
    # via celeste-staff-meal (pyproject.toml)
